adds more review comment stuff
