+++ src/test/java/com/iota/iri/network/pipeline/TransactionProcessingPipelineTest.java
@@ -5,12 +5,10 @@ import com.iota.iri.conf.NodeConfig;
 import com.iota.iri.controllers.TipsViewModel;
 import com.iota.iri.network.NeighborRouter;
 import com.iota.iri.network.SampleTransaction;
-import com.iota.iri.network.TransactionRequester;
 import com.iota.iri.network.neighbor.Neighbor;
 import com.iota.iri.service.milestone.LatestMilestoneTracker;
 import com.iota.iri.service.snapshot.SnapshotProvider;
 import com.iota.iri.storage.Tangle;
-import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -108,7 +106,7 @@ public class TransactionProcessingPipelineTest {
 
     @Test
     public void processingAValidNewTransactionFlowsThroughTheEntirePipeline() throws InterruptedException {
-        TransactionProcessingPipeline pipeline = new TransactionProcessingPipeline();
+        TransactionProcessingPipeline pipeline = new TransactionProcessingPipelineImpl();
         pipeline.init(neighborRouter, nodeConfig, transactionValidator, tangle, snapshotProvider, tipsViewModel,
                 latestMilestoneTracker);
 
@@ -153,7 +151,7 @@ public class TransactionProcessingPipelineTest {
 
     @Test
     public void processingAKnownTransactionOnlyFlowsToTheReplyStage() throws InterruptedException {
-        TransactionProcessingPipeline pipeline = new TransactionProcessingPipeline();
+        TransactionProcessingPipeline pipeline = new TransactionProcessingPipelineImpl();
         pipeline.init(neighborRouter, nodeConfig, transactionValidator, tangle, snapshotProvider, tipsViewModel,
                 latestMilestoneTracker);
 
@@ -187,7 +185,7 @@ public class TransactionProcessingPipelineTest {
     @Test
     public void processingAValidNewTransactionNotOriginatingFromANeighborFlowsThroughTheCorrectStages()
             throws InterruptedException {
-        TransactionProcessingPipeline pipeline = new TransactionProcessingPipeline();
+        TransactionProcessingPipeline pipeline = new TransactionProcessingPipelineImpl();
         pipeline.init(neighborRouter, nodeConfig, transactionValidator, tangle, snapshotProvider, tipsViewModel,
                 latestMilestoneTracker);
 
@@ -233,7 +231,7 @@ public class TransactionProcessingPipelineTest {
 
     @Test
     public void anInvalidNewTransactionStopsBeingProcessedAfterTheValidationStage() throws InterruptedException {
-        TransactionProcessingPipeline pipeline = new TransactionProcessingPipeline();
+        TransactionProcessingPipeline pipeline = new TransactionProcessingPipelineImpl();
         pipeline.init(neighborRouter, nodeConfig, transactionValidator, tangle, snapshotProvider, tipsViewModel,
                 latestMilestoneTracker);
